Improved flight path rendering, some UI refinements

+++ adsb/acprocessor.py
@@ -1,13 +1,16 @@
 from threading import Timer
 from time import sleep
 import datetime
+import logging
 
 from adsb.modesmixer import ModeSMixer
 from adsb.virtualradarserver import VirtualRadarServer
 from adsb.military import MilRanges
 
-from adsb.dbmodels  import Position
+from adsb.db.dbmodels  import Position
 from peewee import IntegrityError
+
+logger = logging.getLogger(__name__)
 
 class AircaftProcessor(object):
 
@@ -28,19 +31,22 @@ class AircaftProcessor(object):
         self._mil_only = config.military_only
         self._db = db
         self.delete_after = config.delete_after
+        # see https://stackoverflow.com/questions/35616602/peewee-operationalerror-too-many-sql-variables-on-upsert-of-only-150-rows-8-c#36788489
+        self._insert_batch_size = 50 
 
     def is_service_alive(self):
         return self._service.connection_alive
 
     def cleanup_items(self):
-        
-        threshold_data = datetime.datetime.utcnow() - datetime.timedelta(minutes=self.delete_after)
-        query = (Position.delete()
-                         .where((Position.timestmp < threshold_data) & (Position.archived == False) ) )
 
-        num_records_deleted = query.execute()
-        if num_records_deleted > 0:
-            print('Deleting {:d} old records from the datbase'.format(num_records_deleted))
+        if self.delete_after > 0:        
+            threshold_data = datetime.datetime.utcnow() - datetime.timedelta(minutes=self.delete_after)
+            query = (Position.delete()
+                            .where((Position.timestmp < threshold_data) & (Position.archived == False) ) )
+
+            num_records_deleted = query.execute()
+            if num_records_deleted > 0:
+                logger.info('Deleting {:d} old records from the datbase'.format(num_records_deleted))
 
     def _run(self):
 
@@ -58,7 +64,8 @@ class AircaftProcessor(object):
             if pos_array:
                 with self._db.atomic():
                     fields = [Position.icao, Position.lat, Position.lon, Position.alt]
-                    Position.insert_many(pos_array, fields=fields).execute()
+                    for i in range(0, len(pos_array), self._insert_batch_size):
+                        Position.insert_many(pos_array[i:i+self._insert_batch_size], fields=fields).execute()
                             
         self.cleanup_items()
